// EduPortal-client/src/app/Models/Graduate.ts
export interface Graduate {
  accountNumber?: string;              
  lastName?: string;                    
  firstName?: string;                 
  institution?: string;                
  idNumber?: string;                   
  mobilePhone?: string;               
  passport?: string;                    

  street?: string;                     
  houseNumber?: string;               
  apartment?: string;                 
  city?: string;                       
  entrance?: string;                    

  fatherCity?: string;                 
  fatherStreet?: string;                
  fatherHouseNumber?: string;          
  fatherEntrance?: string;              
  fatherApartment?: string;            
  mail?: string;                      
  kind?: string;                       
  cycle?: string;                      
  age?: string;                       
  homePhone?: string;                
  addHomePhone?: string;              
  fatherPhone?: string;               
  fatherBusinessPhone?: string;         
  addFatherBusinessPhone?: string;      

  status?: string;                      
}